Extract HttpServerMock
