+++ packages/common-nodejs/src/http/HttpClient.test.ts
@@ -1,64 +1,8 @@
-import { Server } from 'node:http'
-import { assert } from '@marsfoundation/common-universal'
 import { expect } from 'earl'
-import express, { Express } from 'express'
 import { ZodError, z } from 'zod'
 import { Logger } from '../logger/index.js'
 import { HttpClient } from './HttpClient.js'
-
-class HttpServerMock {
-  public app: Express
-  public server: Server | undefined = undefined
-  public requestsCount: Record<string, number>
-
-  constructor() {
-    this.app = express()
-    this.requestsCount = {}
-
-    this.app.use(express.json())
-    this.app.use((req, _res, next) => {
-      this.requestsCount[req.path] = (this.requestsCount[req.path] ?? 0) + 1
-      next()
-    })
-
-    this.app.get('/status', (req, res) => {
-      const status = Number(req.query.status)
-      res.status(status).json({ status })
-    })
-
-    this.app.post('/post', (req, res) => {
-      const bodySafeParse = postSchema.safeParse(req.body)
-      if (!bodySafeParse.success) {
-        res.status(404).end()
-        return
-      }
-      res.status(bodySafeParse.data.status).json(bodySafeParse.data)
-    })
-  }
-
-  listen() {
-    this.server = this.app.listen(0)
-  }
-
-  getUrl(path: string) {
-    const address = this.server?.address()
-    assert(address)
-
-    if (typeof address === 'string') {
-      return address
-    }
-    return `http://127.0.0.1:${address.port}${path}`
-  }
-}
-
-const responseSchema = z.object({
-  status: z.number(),
-})
-
-const postSchema = z.object({
-  status: z.number(),
-})
-type PostSchema = z.infer<typeof postSchema>
+import { HttpServerMock, PostBody, getResponseSchema, postBodySchema } from './HttpServer.mock.js'
 
 describe(HttpClient.name, () => {
   let httpServer: HttpServerMock
@@ -72,13 +16,13 @@ describe(HttpClient.name, () => {
 
   describe('get', () => {
     it('returns successful response', async () => {
-      const httpClient = new HttpClient(new Logger({}))
-      const response = await httpClient.get(httpServer.getUrl('/status?status=200'), responseSchema)
+      const httpClient = new HttpClient(Logger.SILENT)
+      const response = await httpClient.get(httpServer.getUrl('/status?status=200'), getResponseSchema)
       expect(response).toEqual({ status: 200 })
     })
 
     it('throws with invalid schema', async () => {
-      const httpClient = new HttpClient(new Logger({}))
+      const httpClient = new HttpClient(Logger.SILENT)
       const invalidSchema = z.object({
         invalid: z.boolean(),
       })
@@ -88,8 +32,8 @@ describe(HttpClient.name, () => {
     })
 
     it('retries in case of server error', async () => {
-      const httpClient = new HttpClient(new Logger({}))
-      await expect(() => httpClient.get(httpServer.getUrl('/status?status=500'), responseSchema)).toBeRejectedWith(
+      const httpClient = new HttpClient(Logger.SILENT)
+      await expect(() => httpClient.get(httpServer.getUrl('/status?status=500'), getResponseSchema)).toBeRejectedWith(
         'Failed GET: 500 - {"status":500}',
       )
       expect(httpServer.requestsCount['/status']).toEqual(6)
@@ -98,8 +42,8 @@ describe(HttpClient.name, () => {
 
   describe('post', () => {
     it('returns undefined if no schema provided', async () => {
-      const httpClient = new HttpClient(new Logger({}))
-      const body: PostSchema = {
+      const httpClient = new HttpClient(Logger.SILENT)
+      const body: PostBody = {
         status: 200,
       }
 
@@ -107,17 +51,17 @@ describe(HttpClient.name, () => {
     })
 
     it('returns response', async () => {
-      const httpClient = new HttpClient(new Logger({}))
-      const body: PostSchema = {
+      const httpClient = new HttpClient(Logger.SILENT)
+      const body: PostBody = {
         status: 200,
       }
 
-      expect(await httpClient.post(httpServer.getUrl('/post'), body, postSchema)).toEqual(body)
+      expect(await httpClient.post(httpServer.getUrl('/post'), body, postBodySchema)).toEqual(body)
     })
 
     it('throws with invalid schema', async () => {
       const httpClient = new HttpClient(Logger.SILENT)
-      const body: PostSchema = {
+      const body: PostBody = {
         status: 200,
       }
       const invalidSchema = z.object({
@@ -128,8 +72,8 @@ describe(HttpClient.name, () => {
     })
 
     it('retries in case of server error', async () => {
-      const httpClient = new HttpClient(new Logger({}))
-      const body: PostSchema = {
+      const httpClient = new HttpClient(Logger.SILENT)
+      const body: PostBody = {
         status: 500,
       }
       await expect(() => httpClient.post(httpServer.getUrl('/post'), body)).toBeRejectedWith(
